requirements.txt
pydantic
pyserial~=3.5
uvicorn~=0.38.0
dotenv~=0.9.9
python-dotenv==1.2.1
fastapi~=0.124.4